backend/requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==4.2.0
aspose-pdf==24.1.0
astroid==3.0.2
async-timeout==4.0.3
attrs==23.2.0
backoff==2.2.1
beautifulsoup4==4.12.3
blinker==1.4
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
coloredlogs==15.0.1
command-not-found==0.3
contourpy==1.2.0
cryptography==42.0.1
cycler==0.12.1
dataclasses-json==0.6.3
dbus-python==1.2.18
deep-translator==1.11.4
Deprecated==1.2.14
dill==0.3.7
distlib==0.3.7
distro==1.9.0
distro-info==1.1+ubuntu0.2
dnspython==2.5.0
effdet==0.4.1
email-validator==2.1.0.post1
emoji==2.10.0
et-xmlfile==1.1.0
exceptiongroup==1.2.0
faiss-cpu==1.7.4
fastapi==0.109.0
filelock==3.13.1
filetype==1.2.0
flatbuffers==23.5.26
fonttools==4.47.2
frozenlist==1.4.1
fsspec==2023.12.2
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.2
httplib2==0.20.2
httptools==0.6.1
httpx==0.26.0
huggingface-hub==0.20.3
humanfriendly==10.0
idna==3.6
importlib-metadata==4.6.4
iopath==0.1.10
isort==5.13.2
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.3
joblib==1.3.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==2.4
keyring==23.5.0
kiwisolver==1.4.5
langchain==0.1.4
langchain-community==0.0.16
langchain-core==0.1.16
langchain-openai==0.0.5
langdetect==1.0.9
langsmith==0.0.83
launchpadlib==1.10.16
layoutparser==0.3.4
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lxml==5.1.0
Markdown==3.5.2
MarkupSafe==2.1.4
marshmallow==3.20.2
matplotlib==3.8.2
mccabe==0.7.0
more-itertools==8.10.0
mpmath==1.3.0
msg-parser==1.2.0
multidict==6.0.4
mypy-extensions==1.0.0
netifaces==0.11.0
networkx==3.2.1
nltk==3.8.1
numpy==1.26.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.0
olefile==0.47
omegaconf==2.3.0
onnx==1.15.0
onnxruntime==1.15.1
openai==1.10.0
opencv-python==4.9.0.80
openpyxl==3.1.2
orjson==3.9.12
packaging==23.2
pandas==2.2.0
pdf2image==1.17.0
pdfminer.six==20221105
pdfplumber==0.10.3
pikepdf==8.11.2
pillow==10.2.0
platformdirs==4.1.0
portalocker==2.8.2
protobuf==4.25.2
pycairo==1.20.1
pycocotools==2.0.7
pycparser==2.21
pydantic==2.5.3
pydantic-extra-types==2.4.1
pydantic-settings==2.1.0
pydantic_core==2.14.6
PyGObject==3.42.1
PyJWT==2.3.0
pylint==3.0.3
pypandoc==1.12
pyparsing==3.1.1
pypdf==4.0.0
pypdfium2==4.26.0
pytesseract==0.3.10
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.2
python-docx==1.1.0
python-dotenv==1.0.1
python-iso639==2024.1.2
python-magic==0.4.27
python-multipart==0.0.6
python-pptx==0.6.23
pytz==2023.3.post1
PyYAML==6.0.1
rapidfuzz==3.6.1
regex==2023.12.25
requests==2.31.0
safetensors==0.4.2
scipy==1.12.0
SecretStorage==3.3.1
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.25
starlette==0.35.1
sympy==1.12
systemd-python==234
tabulate==0.9.0
tenacity==8.2.3
tiktoken==0.5.2
timm==0.9.12
tokenizers==0.15.1
tomli==2.0.1
tomlkit==0.12.3
torch==2.1.2
torchvision==0.16.2
tqdm==4.66.1
transformers==4.37.1
triton==2.1.0
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
ubuntu-advantage-tools==8001
ufw==0.36.1
ujson==5.9.0
unattended-upgrades==0.1
unstructured==0.12.2
unstructured-client==0.16.0
unstructured-inference==0.7.23
unstructured.pytesseract==0.3.12
urllib3==2.1.0
uvicorn==0.27.0
uvloop==0.19.0
virtualenv==20.24.6
wadllib==1.3.6
watchfiles==0.21.0
websockets==12.0
wrapt==1.16.0
xlrd==2.0.1
XlsxWriter==3.1.9
yarl==1.9.4
zipp==1.0.0
